update:index와 파일디스크립터 분리
<description>
인덱스 스테틱 변수와 파일디스크립터 스테틱 변수 분리
<issue-number>
